Redirect leftover Notion admin pages to AppFlowy.
CMS types and static fallbacks now live in cms-static so AppFlowy adapters no longer import Notion modules.
