+++ src/app/[locale]/admin/notion/page.tsx
@@ -1,5 +1,6 @@
-/**
- * Legacy /admin/notion → AppFlowy contact submissions.
- * Keep the path so bookmarks/nav don't 404 while CMS is AppFlowy-only.
- */
-export { default } from "../appflowy/page";
+import { redirect } from "@/i18n/navigation";
+
+/** Legacy /admin/notion → AppFlowy submissions. */
+export default function LegacyNotionIndexPage() {
+  redirect("/admin/appflowy");
+}
