+++ src/app/[locale]/admin/notion/submissions/page.tsx
@@ -1,2 +1,6 @@
-/** Legacy path — AppFlowy submissions UI. */
-export { default } from "../../appflowy/submissions/page";
+import { redirect } from "@/i18n/navigation";
+
+/** Legacy /admin/notion/submissions → AppFlowy submissions. */
+export default function LegacyNotionSubmissionsPage() {
+  redirect("/admin/appflowy/submissions");
+}
