+++ src/app/[locale]/admin/notion/tasks/page.tsx
@@ -1,2 +1,6 @@
-/** Legacy path — AppFlowy tasks UI. */
-export { default } from "../../appflowy/tasks/page";
+import { redirect } from "@/i18n/navigation";
+
+/** Legacy /admin/notion/tasks → AppFlowy tasks. */
+export default function LegacyNotionTasksPage() {
+  redirect("/admin/appflowy/tasks");
+}
