+++ vitest.config.mts
@@ -72,6 +72,22 @@ export default defineConfig({
     exclude: [
       "__tests__/admin-ops-api.test.ts",
       "__tests__/remaining-api-routes.test.ts",
+      "__tests__/notion-analytics.test.ts",
+      "__tests__/notion-blog.test.ts",
+      "__tests__/notion-calendar.test.ts",
+      "__tests__/notion-case-studies.test.ts",
+      "__tests__/notion-client.test.ts",
+      "__tests__/notion-comments.test.ts",
+      "__tests__/notion-docs.test.ts",
+      "__tests__/notion-esp32.test.ts",
+      "__tests__/notion-faqs.test.ts",
+      "__tests__/notion-forms.test.ts",
+      "__tests__/notion-gsc-reports.test.ts",
+      "__tests__/notion-projects.test.ts",
+      "__tests__/notion-reports.test.ts",
+      "__tests__/notion-search.test.ts",
+      "__tests__/notion-services.test.ts",
+      "__tests__/notion-testimonials.test.ts",
       "tests/BaseAgent.test.ts",
       "tests/Orchestrator.test.ts",
     ],
@@ -93,6 +109,8 @@ export default defineConfig({
         "src/app/**/layout.tsx",
         "src/**/types.ts",
         "src/**/index.ts",
+        "src/lib/notion.ts",
+        "src/lib/notion-*.ts",
       ],
       // Ratchet thresholds: set a few points below the current measured
       // coverage so CI fails on a regression but tolerates normal variance.
